EE-7304 Add self assessment summary to the hmrc domain
