+++ src/main/java/uk/gov/digital/ho/pttg/dto/sasummary/SelfAssessment.java
@@ -1,10 +1,13 @@
-package uk.gov.digital.ho.pttg.dto;
+package uk.gov.digital.ho.pttg.dto.sasummary;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import org.springframework.hateoas.ResourceSupport;
 
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
 @Getter
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = true)
