+++ src/main/java/uk/gov/digital/ho/pttg/dto/sasummary/TaxReturns.java
@@ -1,4 +1,4 @@
-package uk.gov.digital.ho.pttg.dto;
+package uk.gov.digital.ho.pttg.dto.sasummary;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
@@ -6,11 +6,12 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import org.springframework.hateoas.ResourceSupport;
 
+import java.util.List;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @Getter
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = true)
-public class SelfEmployments extends ResourceSupport {
-    private final TaxReturns selfAssessment;
+public class TaxReturns extends ResourceSupport {
+    private final List<TaxReturn> taxReturns;
 }
