+++ src/main/java/uk/gov/digital/ho/pttg/dto/selfemployment/SelfAssessment.java
@@ -1,0 +1,16 @@
+package uk.gov.digital.ho.pttg.dto.selfemployment;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import org.springframework.hateoas.ResourceSupport;
+
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@Getter
+@AllArgsConstructor
+@EqualsAndHashCode(callSuper = true)
+public class SelfAssessment extends ResourceSupport {
+    private final TaxReturns selfAssessment;
+}
