+++ src/main/java/uk/gov/digital/ho/pttg/dto/selfemployment/SelfEmployment.java
@@ -1,4 +1,4 @@
-package uk.gov.digital.ho.pttg.dto;
+package uk.gov.digital.ho.pttg.dto.selfemployment;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
