+++ src/main/java/uk/gov/digital/ho/pttg/dto/selfemployment/TaxReturn.java
@@ -1,0 +1,17 @@
+package uk.gov.digital.ho.pttg.dto.selfemployment;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+
+import java.util.List;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@Getter
+@AllArgsConstructor
+@EqualsAndHashCode
+public class TaxReturn {
+    private final String taxYear;
+    private final List<SelfEmployment> selfEmployments;
+}
