+++ src/main/java/uk/gov/digital/ho/pttg/dto/selfemployment/TaxReturns.java
@@ -1,17 +1,17 @@
-package uk.gov.digital.ho.pttg.dto;
+package uk.gov.digital.ho.pttg.dto.selfemployment;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
+import org.springframework.hateoas.ResourceSupport;
 
 import java.util.List;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @Getter
 @AllArgsConstructor
-@EqualsAndHashCode
-public class TaxReturn {
-    private final String taxYear;
-    private final List<SelfEmployment> selfEmployments;
+@EqualsAndHashCode(callSuper = true)
+public class TaxReturns extends ResourceSupport {
+    private final List<TaxReturn> taxReturns;
 }
